Add files via upload
Files that are in "finalna_verzija" are the files for QtCreator. In there I have made my app for a hydroelectric power plant - basically a control room for the plant. 

The file "finalni_miC.py" is there to act as the dam of the system that is the hydroelectric power plant.

The project uses MQTT to communicate between the dam and the control room.